Refactor and 404 page nav handling
