+++ src/app/components/views/index.js
@@ -3,4 +3,5 @@
 // return
 export { default as About } from "./About"
 export { default as SignIn } from "./SignIn"
-export { default as NotFound } from "./NotFound"
+export { default as Composer } from "./Composer"
+export { default as Submit } from "./Submit"
